Session06

+++ Route.C41.G01.PL/Mapper Helper/DocumentSettings.cs
@@ -2,13 +2,14 @@
 using Microsoft.CodeAnalysis;
 using System;
 using System.IO;
+using System.Threading.Tasks;
 //using static System.Net.WebRequestMethods;
 
 namespace Route.C41.G01.PL.Mapper_Helper
 {
     public static class DocumentSettings
     {
-        public static string UploadFiles(IFormFile file,string folderName )
+        public static async Task<string> UploadFilesAsync(IFormFile file,string folderName )
         {
             //1.Get Located Folder Path
             string FolderPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\Files", folderName);
@@ -29,10 +30,13 @@ namespace Route.C41.G01.PL.Mapper_Helper
          
         public static void DeleteFile(string fileName , string folderName ) 
         {
-            string filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\Files", folderName, fileName);
-            if (File.Exists(filePath) )
+            if (fileName is not null && folderName is not null)
             {
-                File.Delete(filePath);
+                string filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\Files", folderName, fileName);
+                if (File.Exists(filePath))
+                {
+                    File.Delete(filePath);
+                }
             }
 
         }
